+++ CRM/DataprocessorSearch/ActivitySearch.php
@@ -199,4 +199,26 @@ class CRM_DataprocessorSearch_ActivitySearch implements UIFormOutputInterface {
     ));
   }
 
+  /**
+   * Returns the data processor name from a url.
+   *
+   * @param $url
+   *
+   * @return string
+   */
+  public function getDataProcessorNameFromUrl($url) {
+    $dataProcessorName = str_replace('civicrm/dataprocessor_activity_search/', '', $url);
+    return $dataProcessorName;
+  }
+
+  /**
+   * Returns the name of the contact id field.
+   *
+   * @param $config
+   * @return string
+   */
+  public function getContactIdFieldNameFromConfig($config) {
+    return false;
+  }
+
 }