+++ CRM/DataprocessorSearch/Search.php
@@ -209,4 +209,26 @@ class CRM_DataprocessorSearch_Search implements UIFormOutputInterface {
     ));
   }
 
+  /**
+   * Returns the data processor name from a url.
+   *
+   * @param $url
+   *
+   * @return string
+   */
+  public function getDataProcessorNameFromUrl($url) {
+    $dataProcessorName = str_replace('civicrm/dataprocessor_search/', '', $url);
+    return $dataProcessorName;
+  }
+
+  /**
+   * Returns the name of the contact id field.
+   *
+   * @param $config
+   * @return string
+   */
+  public function getContactIdFieldNameFromConfig($config) {
+    return false;
+  }
+
 }